Fix compatibility with np 1.12+

+++ book_figures/chapter10/fig_FFT_sampling.py
@@ -57,7 +57,7 @@ window[i] = 1
 
 #------------------------------------------------------------
 # Compute PSDs
-Nfreq = Nbins / 2
+Nfreq = int(Nbins / 2)
 
 dt = t[1] - t[0]
 df = 1. / (Nbins * dt)
